design: 설문 생성 UI 수정
- height
